+++ lib/screens/home/home_screen.dart
@@ -1,5 +1,8 @@
+import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:hikersafrique/models/event.dart';
+import 'package:hikersafrique/screens/home/homepages/events_page.dart';
+import 'package:hikersafrique/screens/home/homepages/favorites.dart';
 import 'package:hikersafrique/services/auth_notifier.dart';
 import 'package:hikersafrique/services/database.dart';
 import 'package:hikersafrique/items/best_places.dart';
@@ -21,6 +24,19 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
+    int _selectedIndex = 0;
+
+  // Selected item function
+  void onTappedItem(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  List pages = [
+    EventsPage(),
+    Favorites(),
+  ];
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -28,90 +44,31 @@ class _HomeScreenState extends State<HomeScreen> {
         preferredSize: Size.fromHeight(90.0),
         child: CustomHomeAppBar(),
       ),
-      bottomNavigationBar: const CustomBottomNavBar(),
-      body: SafeArea(
-        child: RefreshIndicator(
-          onRefresh: () => Future(() => setState(() {})),
-          child: SingleChildScrollView(
-            child: Column(
-              children: [
-                Row(
-                  children: [
-                    Expanded(
-                      child: SizedBox(
-                        height: 200.0,
-                        child: FutureBuilder<List<Event>>(
-                            future: Database.getSavedEvents(
-                                Provider.of<AuthNotifier>(context)
-                                    .user!
-                                    .clientEmail),
-                            initialData: const [],
-                            builder: (context, snapshot) {
-                              return ListView.builder(
-                                padding: const EdgeInsets.only(right: 20),
-                                itemCount: snapshot.data!.length,
-                                scrollDirection: Axis.horizontal,
-                                shrinkWrap: true,
-                                itemBuilder: (context, int index) {
-                                  return SavedEvent(
-                                    event: snapshot.data![index],
-                                  );
-                                },
-                              );
-                            }),
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(
-                  height: 20.0,
-                ),
-                // Category of best places,most visited etc
-                const SingleChildScrollView(
-                  scrollDirection: Axis.horizontal,
-                  child: Padding(
-                    padding: EdgeInsets.all(10.0),
-                    child: Row(
-                      children: [
-                        // BestPlaces Container
-                        BestPlacesContainer(),
-                        // Most Visited Container
-                        MostVisitedContainer(),
-                        //Favourites Container
-                        FavouriteContainer(),
-                        // New Added Container
-                        NewAddedContainer(),
-                        // Hotels Container
-                        HotelsContainer(),
-                        // Restaurants Container
-                        RestaurantContainer(),
-                      ],
-                    ),
-                  ),
-                ),
-                const SizedBox(
-                  height: 10.0,
-                ),
-                FutureBuilder<List<Event>>(
-                    future: Database.getAvailableEvents(),
-                    initialData: const [],
-                    builder: (context, snapshot) {
-                      return ListView.builder(
-                        itemCount: snapshot.data!.length,
-                        physics: const NeverScrollableScrollPhysics(),
-                        shrinkWrap: true,
-                        itemBuilder: (BuildContext context, index) {
-                          return EventItem(
-                            event: snapshot.data![index],
-                          );
-                        },
-                      );
-                    }),
-              ],
-            ),
-          ),
-        ),
+      drawer: Drawer(
+        
       ),
+      bottomNavigationBar: CurvedNavigationBar(
+      color: Colors.white,
+      animationCurve: Curves.bounceOut,
+      animationDuration: const Duration(milliseconds: 700),
+      buttonBackgroundColor: Colors.white,
+      backgroundColor: Colors.transparent,
+      onTap: (i)=> onTappedItem(i),
+      index: _selectedIndex,
+      items: const [
+        Icon(
+          Icons.person_outlined,
+          size: 30.0,
+        ),
+        Icon(
+          Icons.favorite_outline,
+          size: 30.0,
+        ),
+
+      ],
+    ),
+      body: pages[ _selectedIndex],
+
     );
   }
 }
