+++ lib/screens/home/homepages/favorites.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Favorites extends StatefulWidget {
+  const Favorites({super.key});
+
+  @override
+  State<Favorites> createState() => _FavoritesState();
+}
+
+class _FavoritesState extends State<Favorites> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(child: Text("Favorites page")),
+    );
+  }
+}
