+++ lib/widgets/home_appbar.dart
@@ -15,7 +15,9 @@ class CustomHomeAppBar extends StatelessWidget {
           children: [
             // Sort
             InkWell(
-              onTap: () {},
+              onTap: () {
+                Scaffold.of(context).openDrawer();
+              },
               child: Container(
                 padding: const EdgeInsets.all(10.0),
                 decoration: BoxDecoration(
